chore: Update dependencies and improve logging
This commit updates the project dependencies and adds logging functionality to the SerialAdapter class. The changes aim to improve the overall functionality and reliability of the application. Additionally, unnecessary files and configurations have been removed for a cleaner codebase.

Additionally, this updates to a headless mode, where the jar file can be called using '--auto --headless' which will allow Moonlight devices to use the app - Moonlight native devices that are accepted by Windows as an Xbox 360 controller require a GUI to focus input. Headless mode enables that.